+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-console: ability to filter Ansible hosts by their inventory
+group;
 - concord-server: RBAC checks for process kill and status update
 operations.